(PUP-1678) Provide valid, standard source URIs
The URIs that were used before for the source information were
completely made up. The schema wasn't registered and the path elements
didn't mean much. This replaces it with a scheme where the directory
environments are identified by a file URI that points to the directory,
the legacy environments are identified by a file URI that points to the
puppet configuration file, and any static environments are identified by
a data URI with the content of "internal".