(PUP-1751) Fix service provider on Suse to handle boot services.
The current implementation passes the '--check' flag to chkconfig for
Suse OS family. This causes chkconfig to exit with 0 when the service is
enabled and non-zero when the service is disabled or non-existent.

However, for boot services (services that start with 'boot.'), chkconfig
is returning 0 even if the service is disabled or non-existent.

The fix is to stop passing the '--check' flag to chkconfig and instead
parse the output looking for '<service_name>  on' to see if the service
is enabled.