(PUP-876) Skip resource service check on EL 6
Redmine issues (#4123) and (#4124) added acceptance tests on RHEL
systems to ensure that the services listed by chkconfig matched the list
of services returned by `puppet resource service`, but this assumed that
chkconfig would be authoritative for all running services. Commit
7bdcbea enabled the upstart service provider on EL 6 systems which
breaks this assumption and causes the acceptance tests to fail. This
commit disables these tests on EL 6 since the test is no longer
meaningful.