(maint) Fix useradd spec when using ruby-shadow 2.3.3.
The ruby-shadow gem inserted a new member into the PasswdEntry struct in 2.3.3.
This caused a useradd provider spec to fail that was relying on a member that came after the newly inserted member.
The failure only occurs when the ruby-shadow gem is installed.

The fix is to populate the struct members by name rather than relying on the member order.