(PUP-1563) Don't recurse when dependencies are empty
Previously we would recurse for each leaf dependency unnecessarily.

This commit ensures we only recurse if we have a non-empty set of
dependencies, and adds a debug message when conflicts are checked.

It also adds a spec tests to ensure we evaluate each module for
conflicts only once.

Paired-with: Josh Partlow <joshua.partlow@puppetlabs.com>