(maint) Change infer_set of Array to produce PTupleType
The infer_set was written before PTupleType was implemented and
used an ArrayType with a single PVariantType with the convention
that the PVariant described a sequence. This was a crutch and
not semantically correct. Only code that knew that the result came
from an infer_set could make this assumption about the PVariantType's
content.

The change is to instead use a PTupleType which correctly describes
a sequence of the types in an Array.