(maint) Add method size_range, and repeat_last_range to collections
(resolve conflict from cherry pick to stable and merge to master
back into the implementation work for PUP-1799)

For PCollectionType (PArrayType and PHashType) the method size_range
now returns a range [min, max] count of entries accepted in the
collection. A result is always given even if no size constraint has
been set.

For PTupleType, the method size_range gives the total accepted
parameter range, and the method repeat_last_range returns the range
of the last parameter. For a tuple, if the size is not constrained
the repeat_last_range is always [1,1]. For a tuple type, the
range of the last parameter is considered in the size_range.

These methods helps answering questions about these types without having
to know anything about their internal semantics.