(PUP-1596) Fix MemoryFile to string conversion in Ruby 1.8
There were two expectations in the previous PUP-1596 work; that a
to_path would permit File functions to operate on a MemoryFile as if it
were a Pathname like thing, and that String#clear was present (in the
spec unit).  A to_str is provided on MemoryFile now for a
File.expand_path and ilk on Ruby 1.8.7, and we're no longer trying to
clear the :content String in the environments_spec.