(PUP-1799) Make new Function API report parameter names.
This makes the function API include the names of parameters if
they are available. When names are available the signatures
show the names (for both array and tuple based signatures).

This also handles the changes made to the TupleType (size constraint
is for the entire type, not just the last entry).