(maint) Fix directoryservice spec order issue.
The directoryservice provider caches some information such as OS version
and groups.  If a spec runs before the directoryservice spec that causes
information to be cached, expected stubs in the directoryservice spec
will not be invoked.

The fix is to remove the instance variables that are caching the
information on the examples that are expecting that nothing is already
cached.