(maint) Check for files by full name
Periodically this test would fail with a message about some path being a
directory. It turns out that what was happening was that that generated
name for the temp directory would end in "one" or "two". This caused the
test to detect the directory entry in the response from the search
request to be one of the files and so it would try to read the contents,
which wasn't possible because it was the directory.

This changes it to check for the files by the full path and use
"include_in_any_order" to pull together two tests into one.