(PUP-1799) Add marking of optional and varargs count in error message
When an error is raised becuse of no matching signature, the 
expected signatures are printed with each parameter on the form:

Type name<count>

where count is ? for an optional {0,1} parameter, and with a range
{from, to} if from and to are not both 1 (in which case count is an 
empty string), and only a comma is shown if to is INFINITY. 

e.g. a message may show:

  foo(Integer a, String b?, String{0,}) - arg count {1,}

The trailing arg count shows required min, and allowed max as a short
hand notation.