(PUP-1799) Add support for injection of parameters and weaving
This adds injection of parameters and weaving of them when
dispatching. The dispatching logic is also somewhat refactored
from using anonymous arrays to a Dispatch object.

The methods injected_param, injected_producer_param specifies
that additional parameters should be injected when a call is made.
Their position defines where they are injected and the ruby method
signature should be matched to handle them.

The injection of parameters works for both regular and polymorphic
dispatch.

This makes use of injection more declarative, and it is easy to
test a Function class' methods.