(maint) Fix package acceptance test failure caused by gem provider changes.
The gem provider recently had the install_options feature implemented.
The provider was appending a nil value to the command arguments if there
are no install_options specified. This gets converted to an empty string
before the Kernel.exec call and causes gem to treat the empty string
as a gem name. The result is a failure when trying to install gems in
the acceptance test.

The fix is to append install_options only if specified on the resource.