(maint) Update gem provider spec for recent change.
The spec had an example which was expecting a nil
install_options by checking the 6th element in an array.
With the change to not append the install_options
unless specified, the example was passing but the
functionality no longer matched the example.

Changing the example to ensure 5 elements were
passed to execute instead of the previous 6.