(PUP-1799) Fix test of function API for Ruby 1.8.7
The tests when running under Ruby 1.8.7 have to be different since
the parameter names are not available in the API for method. The
generated error messages therefore looks different. Also, since the
parameter arity is quite broken in Ruby 1.8.7 it is not possible to
represent a signature with a good type - the result is that an
PArrayType is used to represent the signature, and it is always
Object{to, from}.

The change is to compare against different messages in the 3 test cases
that failed on Ruby 1.8.7.