(PUP-1433) Stop deprecating env settings on CLI
When config_version, modulepath, and manifest were specified on the
command line, the code that updated the current_environment with the
information looked it up in the settings. This changes it to use the
value() method which isn't going to issue the warning. After the
deprecation period is over the command line handling will be changed to
just global application command line handlers and the environment
configuration will be driven from that.