(PUP-485) Change format for actual type in assert_type error message
Instead of showing all details about the inferred type of the
given value, it is now made generic before output. This is good
for the simpler cases where it is confusing to see e.g. Integer[11,11]
when the value was 11 (an exact range). It does however mean that
information is "lost" when matching against more complex types. A more
complex "type diff" may be required to provide really good output.
Another alternative is to just display the value (not the type) of the
actual - but that may be very long if it is an Array or Hash.