(PUP-2035) (PUP-1799) Use new 4x loaders for 4x functions
This wires the new loaders (via the Puppet context) into the future
evaluator to make it lookup 4x functions via the new loader
infrastructure.

Currently, there is one function 'assert_type' available to be
called.

When the option Puppet[:biff] is true, all 3x functions are turned
into 4x functions. This is not fully operational yet, only functions
that do not rely on 3x scope features can be called that way.